Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
qdrant-client==1.6.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
black==23.11.0
sqlalchemy[asyncio]==2.0.23
asyncpg
numpy==1.26.2
markdown==3.5.1
fastembed>=0.3.0
onnxruntime>=1.16.0
google-generativeai==0.4.0